+++ projects/kinara-common/src/hashfuncs/MurmurHash.cpp
@@ -247,7 +247,7 @@ void MurmurHash3_x64_128(const void* key, const u64 len,
                          const uint32_t seed, void* out)
 {
     const uint8_t * data = static_cast<const u08*>(key);
-    const int nblocks = len / 16;
+    const u64 nblocks = len / 16;
 
     uint64_t h1 = seed;
     uint64_t h2 = seed;
